requirements.txt
asgiref==3.7.2
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.4.0
cryptography==42.0.5
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.10
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==1.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sqlparse==0.4.4
urllib3==2.2.3
